+++ engine/lite/components/request-response/src/main/scala/pl/touk/nussknacker/engine/lite/components/requestresponse/jsonschema/sources/JsonSchemaRequestResponseSourceFactory.scala
@@ -9,25 +9,27 @@ import pl.touk.nussknacker.engine.api.typed._
 import pl.touk.nussknacker.engine.api.typed.typing.Unknown
 import pl.touk.nussknacker.engine.api.{MetaData, NodeId}
 import pl.touk.nussknacker.engine.json.{JsonSchemaExtractor, SwaggerBasedJsonSchemaTypeDefinitionExtractor}
-import pl.touk.nussknacker.engine.requestresponse.api.openapi.RequestResponseOpenApiSettings.InputSchemaProperty
+import pl.touk.nussknacker.engine.requestresponse.api.openapi.RequestResponseOpenApiSettings.{InputSchemaProperty, OutputSchemaProperty}
 import pl.touk.nussknacker.engine.requestresponse.api.{RequestResponseSource, RequestResponseSourceFactory}
 
 class JsonSchemaRequestResponseSourceFactory extends RequestResponseSourceFactory with SingleInputGenericNodeTransformation[Source] {
 
-  override type State = Schema
+  override type State = (Schema, Schema)
 
   private val nodeIdDependency = TypedNodeDependency[NodeId]
   private val metaDataDependency = TypedNodeDependency[MetaData]
+
   override def nodeDependencies: List[NodeDependency] = List(nodeIdDependency, metaDataDependency)
 
   private val jsonSchemaExtractor = new JsonSchemaExtractor()
 
   override def contextTransformation(context: ValidationContext, dependencies: List[NodeDependencyValue])(implicit nodeId: NodeId): NodeTransformationDefinition = {
     case TransformationStep(Nil, _) =>
-      val determinedSchema = jsonSchemaExtractor.getSchemaFromProperty(InputSchemaProperty, dependencies)
-      val validationResult = determinedSchema.swap.toList.flatMap(_.toList)
-      val finalState = determinedSchema.toOption
-      val finalInitializer = determinedSchema.toOption.fold(new BasicContextInitializer(Unknown)) { schema =>
+      val inputSchema = jsonSchemaExtractor.getSchemaFromProperty(InputSchemaProperty, dependencies)
+      val outputSchema = jsonSchemaExtractor.getSchemaFromProperty(OutputSchemaProperty, dependencies)
+      val validationResult = inputSchema.product(outputSchema).swap.toList.flatMap(_.toList)
+      val finalState = inputSchema.product(outputSchema).toOption
+      val finalInitializer = inputSchema.toOption.fold(new BasicContextInitializer(Unknown)) { schema =>
         val schemaTypingResult = SwaggerBasedJsonSchemaTypeDefinitionExtractor.swaggerType(schema).typingResult
         new BasicContextInitializer(schemaTypingResult)
       }
@@ -36,11 +38,11 @@ class JsonSchemaRequestResponseSourceFactory extends RequestResponseSourceFactor
 
   override def implementation(params: Map[String, Any],
                               dependencies: List[NodeDependencyValue],
-                              finalStateOpt: Option[Schema]): RequestResponseSource[Any] = {
-    val finalSchemaState = finalStateOpt.getOrElse(throw new IllegalStateException("Unexpected (not defined) final state determined during parameters validation"))
+                              finalStateOpt: Option[(Schema, Schema)]): RequestResponseSource[Any] = {
+    val (inputSchemaState, outputSchemaState) = finalStateOpt.getOrElse(throw new IllegalStateException("Unexpected (not defined) final state determined during parameters validation"))
     val nodeId = nodeIdDependency.extract(dependencies)
     val metaData = metaDataDependency.extract(dependencies)
-    new JsonSchemaRequestResponseSource(finalSchemaState.toString, metaData, finalSchemaState, nodeId)
+    new JsonSchemaRequestResponseSource(inputSchemaState.toString, metaData, inputSchemaState, outputSchemaState, nodeId)
   }
 
 }